task/us99-objects-schema-add-locations
The Objects schema has a .location. item that is a dictionary
construct that contains latitude and longitude details and is currently missing from the schema.

Turns out that this is due to variations provided in the sample data for sprint 2.  Not going to chase this until project.s future is clarified

oct_schema/model_base.py
* changed _batch_loader() into a @classmethod as the query() was creating a new, blank Document object